requirements.txt
streamlit==1.27.2
pandas==2.1.1
plotly==5.17.0
Pillow==10.0.1
urllib3==2.0.7
numpy==1.24.3
python-dotenv==1.0.0
bcrypt==4.0.1
matplotlib==3.7.2
seaborn==0.12.2 